Version 1.0
Config::IniMan - INI file manager object style preserved format and sort
order

This module reads and writes INI files in object style and preserves
original files sort order, comments, empty lines, and multi lines
parameters.

It is basically built on the top of using the Tie::IxHash module which
implements Perl hashes that preserve the order in which the hash
elements were added. The order is not affected when values corresponding
to existing sections or parameters are changed.

New sections will be added to the end of the current file contents and
new parameters will be added to the end of the current section.